reader-service-py/requirements.txt
nats-py==2.10.0
influxdb-client==1.49.0
pandas==2.3.0